add navbar to repo
