+++ src/App.js
@@ -1,19 +1,18 @@
 import React from 'react';
-import { Route, Routes, BrowserRouter } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from './store';
 import Nav from './components/Nav';
+import './app.scss';
 
 const App = () => (
-  <>
+  <div className="app">
     <BrowserRouter>
       <Provider store={store}>
-        <Routes>
-          <Route path="/" element={<Nav />} />
-        </Routes>
+        <Nav />
       </Provider>
     </BrowserRouter>
-  </>
+  </div>
 );
 
 export default App;
