+++ src/app.scss
@@ -1,0 +1,4 @@
+.app {
+  display: flex;
+  flex-direction: row;
+}
