--- src/nav.css
@@ -1,7 +1,0 @@
-.split {
-  display: flex;
-}
-
-ul {
-  border-right: 2px solid black;
-}
