+++ src/styles/home.scss
@@ -1,0 +1,4 @@
+.homeContainer {
+  width: 80%;
+  text-align: center;
+}
