+++ src/styles/navbar.scss
@@ -1,0 +1,44 @@
+.navbar {
+  display: flex;
+  flex-direction: column;
+  margin: 0 1px;
+  text-align: center;
+  height: 96vh;
+  width: 200px;
+
+  img {
+    position: relative;
+    left: 10px;
+    margin: 0 auto;
+    margin-bottom: 50px;
+    width: 150px;
+    height: 150px;
+    object-fit: cover;
+    border-radius: 50%;
+  }
+
+  a {
+    border-radius: 4px;
+    padding: 6px 10px;
+    width: 100%;
+    text-decoration: none;
+    margin: 5px 0;
+    font-weight: 600;
+    font-size: 24px;
+  }
+  
+  a.active {
+    background-color: greenyellow;
+    color: white;
+  }
+  
+  a.inactive {
+    color: black;
+  }
+}
+
+hr {
+  position: relative;
+  left: 30px;
+  margin: 0 50px 0 0;
+}
